Wflow

This is an in development version of wflow, a continuation of the work available here. For now using this for any real calculations is not advised, please use the linked stable wflow release for that.

Wflow.pcrdir

Wflow.Config

Wflow.active_indices

Wflow.acttransp_unsat_sbm

Wflow.add_time

Wflow.close_files

Wflow.csv_header

Wflow.detdrainlength

Wflow.detdrainwidth

Wflow.equal_size_vectors

Wflow.flat!

Wflow.flowgraph

Wflow.fraction_runoff_toriver

Wflow.get_at!

Wflow.glacier_hbv

Wflow.infiltration

Wflow.initialize_hbv_model

Wflow.initialize_sbm_model

Wflow.initialize_storage

Wflow.kin_wave!

Wflow.kinematic_wave

Wflow.ncread

Wflow.param

Wflow.pow

Wflow.rainfall_interception_gash

Wflow.rainfall_interception_modrut

Wflow.reducer

Wflow.reducerfunction

Wflow.run_simulation

Wflow.scurve

Wflow.set_layerthickness

Wflow.set_states

Wflow.setup_netcdf

Wflow.snowpack_hbv

Wflow.statevars

Wflow.sum_at

Wflow.symbols

Wflow.timecycles

Wflow.unsatzone_flow_layer

Wflow.unsatzone_flow_sbm

Wflow.update

Wflow.update

Wflow.update_cyclic!

Wflow.update_forcing!

Wflow.write_output

Wflow.@symbols_str

Wflow.pcrdir — Constant

Map from PCRaster LDD value to a CartesianIndex

Wflow.Config — Type

Parsed TOML configuration

Wflow.active_indices — Method

active_indices(subcatch_2d, nodata)

Takes a 2D array of the subcatchments. And derive forward and reverse indices.

1: Get a list of CartesianIndex{2}` that are active, based on a nodata value. These map from the 1D internal domain to the 2D external domain.

2: Make a reverse index, a Matrix{Int}`, which maps from the 2D external domain to the 1D internal domain, providing an Int which can be used as a linear index. Values of 0 represent inactive cells.

Wflow.acttransp_unsat_sbm — Function

acttransp_unsat_sbm(rootingdepth, ustorelayerdepth, sumlayer, restpotevap, sum_actevapustore, c, usl, θₛ, θᵣ, hb, ust::Bool = false)

Compute actual transpiration for unsaturated zone. If ust is true, the whole unsaturated store is available for transpiration.

Arguments

rootingdepth

ustorelayerdepth

sumlayer (depth (z) of upper boundary unsaturated layer)

restpotevap (remaining evaporation)

sum_actevapustore (cumulative actual transpiration (more than one unsaturated layers))

c (Brooks-Corey coefficient)

usl (thickness of unsaturated zone)

θₛ

θᵣ

hb (air entry pressure)

ust

Output

ustorelayerdepth

sum_actevapustore

restpotevap

Wflow.add_time — Method

Add a new time to the unlimited time dimension, and return the index

Wflow.close_files — Method

Close input and output datasets that are opened on model initialization

Wflow.csv_header — Method

Get a Vector{String} of all columns names for the CSV header, exept the first, time

Wflow.detdrainlength — Method

detdrainlength(ldd, xl, yl)

Determines the drainaige length for a non square grid. Input ldd (drainage network), xl (length of cells in x direction), yl (length of cells in y direction). Output is drainage length.

Wflow.detdrainwidth — Method

detdrainwidth(ldd, xl, yl)

Determines the drainaige width for a non square grid. Input ldd (drainage network), xl (length of cells in x direction), yl (length of cells in y direction). Output is drainage width.

Wflow.equal_size_vectors — Method

equal_size_vectors(x)

Used in the structs of arrays to ensure all vectors are of equal length.

equal_size_vectors(([1,2], [1,2,3])) would throw an ArgumentError. equal_size_vectors(([4,5], [4,5])) would pass. equal_size_vectors((1, [4,5], [4,5])) would also pass, since 1 is not an AbstractVector.

Wflow.flat! — Method

Flatten a nested dictionary, keeping track of the full address of the keys. Useful for converting TOML of the format:

[a.b]
field_of_b = "name_in_output"

[a.d.c]
field_of_c = "other_name_in_output"

to a non-nested format:

Dict( symbols"a.b.fieldofb" => "nameinoutput, symbols"a.d.c.fieldofc" => "othernamein_output, )

Wflow.flowgraph — Method

Convert a gridded drainage direction to a directed graph

Wflow.fraction_runoff_toriver — Method

fraction_runoff_toriver(graph, ldd, index_river, slope, n)

Determine ratio frac between slope river cell index_river and slope of each upstream neighbor (based on directed acyclic graph graph).

Wflow.get_at! — Method

Extract a NetCDF variable at a given time

Wflow.glacier_hbv — Method

glacier_hbv(glacierfrac, glacierstore, snow, temperature, tt, cfmax, g_sifrac, Δt)

HBV-light type of glacier modelling. First, a fraction of the snowpack is converted into ice using the HBV-light model (fraction between 0.001-0.005 per day). Glacier melting is modelled using a temperature degree factor and only occurs if the snow cover < 10 mm.

Arguments

glacierFrac fraction covered by glaciers [-]

glacierstore volume of the glacier [mm] w.e.

snow snow pack on top of glacier [mm]

temperature air temperature [°C]

tt temperature threshold for ice melting [°C]

cfmax ice degree-day factor in [mm/(°C/day)]

g_sifrac fraction of the snow turned into ice [-]

Δt model timestep [s]

Output

snow

snow2glacier

glacierstore

glaciermelt

Wflow.infiltration — Method

infiltration(avail_forinfilt, pathfrac, cf_soil, tsoil, infiltcapsoil, infiltcappath, ustorecapacity, modelsnow::Bool, soilinfreduction::Bool)

Soil infiltration based on infiltration capacity soil infiltcapsoil, infiltration capacity compacted area infiltcappath and capacity unsatured zone ustorecapacity. The soil infiltration capacity can be adjusted in case the soil is frozen (modelsnow and soilinfreduction is true).

Wflow.initialize_hbv_model — Method

initialize_hbv_model(config::Config)

Initial part of the SBM model concept. Reads the input settings and data as defined in the Config object. Will return a Model that is ready to run.

Wflow.initialize_sbm_model — Method

initialize_sbm_model(config::Config)

Initial part of the SBM model concept. Reads the input settings and data as defined in the Config object. Will return a Model that is ready to run.

Wflow.initialize_storage — Method

Determine the initial storage depending on the storage function

Wflow.kin_wave! — Method

Kinematic wave flow rate over the whole network for a single timestep

Wflow.kinematic_wave — Method

Kinematic wave flow rate for a single cell and timestep

Wflow.ncread — Method

ncread(nc, var; <keyword arguments>)

Read parameter var from NetCDF file nc. Supports various keyword arguments to get selections of data in desired types, with or without missing values.

Arguments

key=identity: key is a function which maps var to the name of the parameter in the NetCDF file. The default identity keeps var as is.

sel=nothing: a selection of indices, such as a Vector{CartesianIndex} of active cells, to return from the NetCDF. By default all cells are returned.

defaults=nothing: a dictionary in which default values are looked up if var is not in nc. By default it gives an error in this case.

type=nothing: type to convert data to after reading. By default no conversion is done.

allow_missing=false: Missing values within sel is not allowed by default. Set to true to allow missing values.

fill=nothing: Missing values are replaced by this fill value if allow_missing is false.

dimname : name of third dimension of parameter var. By default no third dimension is expected.

Wflow.param — Method

Get a nested field using a tuple of Symbols

Wflow.pow — Method

Faster method for exponentiation

Wflow.rainfall_interception_gash — Method

rainfall_interception_gash(cmax, e_r, canopygapfraction, precipitation, canopystorage;maxevap = 9999.0)

Interception according to the Gash model (for daily timesteps). cmax is the maximum canopy storage and e_r is the ratio of the average evaporation from the wet canopy and the average precipitation intensity on a saturated canopy.

Wflow.rainfall_interception_modrut — Method

rainfall_interception_modrut(precipitation, potential_evaporation, canopystorage, canopygapfraction, cmax)

Interception according to a modified Rutter model. The model is solved explicitly and there is no drainage below cmax.

Wflow.reducer — Method

Get a reducer function based on CSV output settings defined in a dictionary

Wflow.reducerfunction — Method

Mapping from reducer strings in the TOML to functions

Wflow.run_simulation — Method

run_simulation(tomlpath::String)
run_simulation(config::Config)
run_simulation(model::Model)

Run an entire simulation starting either from a path to a TOML settings file, a prepared Config object, or an initialized Model object. This allows more flexibility if you want to for example modify a Config before initializing the Model.

Wflow.scurve — Method

scurve(x; a = 0.0, b = 1.0, c = 1.0)

Sigmoid "S"-shaped curve.

Arguments

x::Real: input

a::Real: determines the centre level (default = 0.0)

b::Real: determines the amplitude of the curve

c::Real: determines the steepness or "stepwiseness" of the curve. The higher c the sharper the function. A negative c reverses the function.

Wflow.set_layerthickness — Method

set_layerthickness(d::Float64, sl::SVector)

Calculate actual soil thickness of layers based on a reference depth (e.g. soil depth or water table depth) d, a SVector sl with cumulative soil depth starting at soil surface (0), and a SVector tl with actual thickness per soil layer.

Wflow.set_states — Method

set_states(instate_path, model, states, sel, config ; <keyword arguments>)

Read states contained in Tuple states from NetCDF file located in instate_path, and set states in model object. Active cells are selected with sel (Vector{CartesianIndex}) from the NetCDF file.

Arguments

type = nothing: type to convert data to after reading. By default no conversion is done.

Wflow.setup_netcdf — Method

prepare an output dataset

Wflow.snowpack_hbv — Method

snowpack_hbv(snow, snowwater, precipitation, temperature, tti, tt, ttm, cfmax, whc)

HBV type snowpack modeling using a temperature degree factor. All correction factors (RFCF and SFCF) are set to 1. The refreezing efficiency factor is set to 0.05.

Arguments

snow (snow storage)

snowwater (liquid water content in the snow pack)

precipitation (throughfall + stemflow)

temperature

tti (snowfall threshold interval length)

tt (threshold temperature for snowfall)

ttm (melting threshold)

cfmax (degree day factor, rate of snowmelt)

whc (Water holding capacity of snow)

rfcf correction factor for rainfall

sfcf correction factor for snowfall

cfr refreeing efficiency constant in refreezing of freewater in snow

Output

snow

snowwater

snowmelt

rainfall (precipitation that occurs as rainfall)

snowfall (precipitation that occurs as snowfall)

Wflow.statevars — Function

statevars(::SBM; snow=false)
statevars(::LateralSSF)
statevars(::SurfaceFlow)
statevars(::NaturalLake)
statevars(::SimpleReservoir)

Get a tuple of symbols representing the fields that are model states.

Wflow.sum_at — Method

Return the sum of the array A at indices inds

Wflow.symbols — Method

Turn "a.aa.aaa" into (:a, :aa, :aaa)

Wflow.timecycles — Method

timecycles(times)

Given a vector of times, return a tuple of (month, day) for each time entry, to use as a cyclic time series that repeats every year. By using monthday rather than dayofyear, leap year offsets are avoided.

It can generate such a series from eiher TimeTypes given that the year is constant, or it will interpret integers as either months or days of year if possible.

Wflow.unsatzone_flow_layer — Method

unsatzone_flow_layer(usd, kv_z, l_sat, c)

Assuming a unit head gradient, the transfer of water from an unsaturated store layer usd is controlled by the vertical saturated hydraulic conductivity kv_z (bottom layer or water table), the effective saturation degree of the layer (ratio usd and l_sat), and a Brooks-Corey power coefficient c.

Wflow.unsatzone_flow_sbm — Method

unsatzone_flow_sbm(ustorelayerdepth, soilwatercapacity, satwaterdepth, kv_z, usl, θₛ, θᵣ)

The transfer of water from the unsaturated store ustorelayerdepth to the saturated store satwaterdepth is controlled by the vertical saturated hydraulic conductivity kv_z at the water table and the ratio between ustorelayerdepth and the saturation deficit (soilwatercapacity minus satwaterdepth). This is the original Topog_SBM vertical transfer formulation.

Wflow.update — Method

Update a single lake at position i.

This is called from within the kinematic wave loop, therefore updating only for a single element rather than all at once.

Wflow.update — Method

Update a single reservoir at position i.

This is called from within the kinematic wave loop, therefore updating only for a single element rather than all at once.

Wflow.update_cyclic! — Method

Get cyclic NetCDF input for the given time

Wflow.update_forcing! — Method

Get dynamic NetCDF input for the given time

Wflow.write_output — Method

Write model output

Wflow.@symbols_str — Macro

Turn symbols"a.aa.aaa" into (:a, :aa, :aaa)